+++ tiledb/sm/query/writers/writer_base.cc
@@ -1203,3 +1203,17 @@ bool WriterBase::remote_query() const {
 }
 
 }  // namespace tiledb::sm
+
+template <>
+IndexedList<tiledb::sm::WriterTileTuple>::IndexedList(
+    shared_ptr<tiledb::sm::MemoryTracker> memory_tracker)
+    : memory_tracker_(memory_tracker)
+    , list_(memory_tracker->get_resource(sm::MemoryType::TILE_WRITER_DATA)) {
+}
+
+template <>
+IndexedList<tiledb::common::IndexedList<tiledb::sm::WriterTileTuple>>::
+    IndexedList(shared_ptr<tiledb::sm::MemoryTracker> memory_tracker)
+    : memory_tracker_(memory_tracker)
+    , list_(memory_tracker->get_resource(sm::MemoryType::TILE_WRITER_DATA)) {
+}
